requirements.txt
zai-sdk
pypdf
pypdfium2
python-dotenv
Pillow
PySide6
